+++ src/app/types/enums.ts
@@ -1,0 +1,11 @@
+export enum ReactionsRange {
+  OVERTIME = 'overtime',
+  WEEKLY = 'weekly',
+  MONTHLY = 'monthly'
+}
+
+export enum BackgroundMode {
+  RANDOM,
+  SINGLE,
+  COLOR
+}
