+++ src/app/types/types.ts
@@ -50,19 +50,6 @@ export interface UserResetConfirm {
   password: string
 }
 
-export interface PaginatedResponse {
-  "meta": {
-    "current_page": number,
-    "from": number,
-    "last_page": number,
-    "path": string,
-    "per_page": number,
-    "to": number,
-    "total": number
-  },
-  "data": {}[]
-}
-
 export interface FavoritesResponse {
   per_page: number
   pages: number
